+++ aldryn_forms/__init__.py
@@ -1,10 +1,2 @@
 # -*- coding: utf-8 -*-
-import warnings
-
-__version__ = '2.0.4'
-
-warnings.warn(
-    'The "aldryn_forms.FormData" model has been deprecated in '
-    'favor of aldryn_forms.FormSubmission',
-    DeprecationWarning
-)
+__version__ = '2.0.4.7'